workers/requirements.txt
rq==1.16.2
redis==5.0.8
psycopg2-binary==2.9.10








